refactor: cmake setup (#23)
* feat:  Initial refactor of cmake

Refactor cmake and project organization to be more modular and to better conform to best practices.

* chore:  Remove unused file

* fix: Correct build errors

Corrected tests link errors by reverting GTest back to hunter package instead of submodule.  Removed Windows compile flags for /MT

* fix: Added missing header

* chore: Moved helpers to private header

* fix:  Made http public header

* chore:  Removed unused headers

* fix:  Remove unnecessary strings

Changed tests to just compare native types instead of converting to strings when not needed.

* fix:  Fix PlatformIO builds

Fixed PlatformIO builds to work with new folder structure.  Removed submodule python script in favor of just ignoring the directory.  This will allow desktop and IoT to better co-exist in the same source tree. Remove [common] section of test/platformio.ini as it did not seem to actually be respected.

* fix:  Remove old submodule

* fix: Correct test binary path

* fix:  Corrected tests

* chore:  Added Windows build files to .gitignore

* fix:  Correct test where blocks can be a zero value

* fix:  Add missing macro

Added missing adapter macro to GTest adapter.

Will submit PR to AUnit library for proper fix.
