+++ .gitmodules
@@ -1,3 +1,6 @@
-[submodule "test/include/ArduinoJson"]
-	path = test/include/ArduinoJson
+[submodule "test/lib/ArduinoJson"]
+	path = test/lib/ArduinoJson
 	url = https://github.com/bblanchon/ArduinoJson
+[submodule "test/lib/googletest"]
+	path = test/lib/googletest
+	url = https://github.com/google/googletest.git
