+++ platformio.ini
@@ -14,7 +14,7 @@
 platform = espressif8266
 board = huzzah
 framework = arduino
-build_flags = -I./src/
+build_flags = -I./src/ -I./src/include/cpp-client
 src_filter = +<*> -<.git/> -<svn/> -<example/> -<examples/> -<http/os/> -<_3rdParty> -<CMakeFiles>
 upload_speed = 921600
 
@@ -22,6 +22,6 @@ upload_speed = 921600
 platform = espressif32
 board = esp32dev
 framework = arduino
-build_flags = -I./src/
+build_flags = -I./src/ -I./src/include/cpp-client
 src_filter = +<*> -<.git/> -<svn/> -<example/> -<examples/> -<http/os/> -<_3rdParty> -<CMakeFiles>
 upload_speed = 921600
