+++ src/include/cpp-client/api/one/delegates/one_delegates.h
@@ -12,7 +12,6 @@
 
 #include "api/api_base.h"
 #include "api/one/one_paths.h"
-#include "helpers/helpers.h"
 
 namespace Ark {
 namespace Client {
