+++ src/include/cpp-client/api/one/transactions/one_transactions.h
@@ -12,7 +12,6 @@
 
 #include "api/api_base.h"
 #include "api/one/one_paths.h"
-#include "helpers/helpers.h"
 #include <string>
 
 namespace Ark {
