+++ src/include/cpp-client/api/two/blocks/two_blocks.h
@@ -7,25 +7,26 @@
  * file that was distributed with this source code.
  **/
 
-#ifndef TWO_VOTES_H
-#define TWO_VOTES_H
+#ifndef TWO_BLOCKS_H
+#define TWO_BLOCKS_H
 
 #include "api/api_base.h"
 #include "api/two/two_paths.h"
-#include "helpers/helpers.h"
 
 namespace Ark {
 namespace Client {
 namespace API {
 namespace TWO {
 
-class Votes : public ApiBase
+class Blocks : public ApiBase
 {
 public:
-  Votes() = default;
+  Blocks() = default;
 
-  std::string get(const char *const identifier);
+  std::string get(const char *const blockId);
   std::string all(int limit = 5, int page = 1);
+  std::string transactions(const char *const blockId);
+  std::string search(std::pair<const char*, const char*> bodyParameters, int limit = 5, int page = 1);
 };
 
 };
