+++ src/include/cpp-client/api/two/votes/two_votes.h
@@ -7,28 +7,24 @@
  * file that was distributed with this source code.
  **/
 
-#ifndef TWO_TRANSACTIONS_H
-#define TWO_TRANSACTIONS_H
+#ifndef TWO_VOTES_H
+#define TWO_VOTES_H
 
 #include "api/api_base.h"
 #include "api/two/two_paths.h"
-#include "helpers/helpers.h"
 
 namespace Ark {
 namespace Client {
 namespace API {
 namespace TWO {
 
-class Transactions : public ApiBase
+class Votes : public ApiBase
 {
 public:
-  Transactions() = default;
+  Votes() = default;
 
-  std::string getUnconfirmed(const char *const identifier, int limit = 2, int page = 1);
-  std::string get(const char *const identifier, int limit = 5, int page = 1);
+  std::string get(const char *const identifier);
   std::string all(int limit = 5, int page = 1);
-  std::string allUnconfirmed(int limit = 2, int page = 1);
-  std::string types(int limit = 5, int page = 1);
 };
 
 };
