+++ src/include/cpp-client/api/two/wallets/two_wallets.h
@@ -12,7 +12,6 @@
 
 #include "api/api_base.h"
 #include "api/two/two_paths.h"
-#include "helpers/helpers.h"
 #include <string>
 
 namespace Ark {
