+++ test/iot/gtest/gtest.h
@@ -6,4 +6,6 @@
 #include <AUnit.h>
 #include <aunit/contrib/gtest.h>
 
+#define ASSERT_GE(e, a) assertMoreOrEqual(static_cast<decltype(a)>(e), a)
+
 #endif
