+++ test/platformio.ini
@@ -12,23 +12,20 @@
 src_dir = ..
 lib_dir = ..
 
-[common]
-lib_deps = ArduinoJson@5.13.2, AUnit
-
 [env:esp8266]
+lib_deps = ArduinoJson, AUnit
 platform = espressif8266
 board = huzzah
 framework = arduino
-build_flags = -I../test/iot/ -DUNIT_TEST
-src_filter = +<*> -<.git/> -<svn/> -<example/> -<examples/> -<bin> -<../src/http/os/> -<_3rdParty> -<CMakeFiles>
-extra_scripts = pre:rm_submodules.py
+build_flags = -I../test/iot/ -I../test -I../src -I../src/include/cpp-client -DUNIT_TEST
+src_filter = +<*> -<.git/> -<svn/> -<example/> -<examples/> -<bin> -<../src/http/os/> -<_3rdParty> -<CMakeFiles> -<test/lib/googletest> -<test/lib/ArduinoJson>
 upload_speed = 921600
 
 [env:esp32]
+lib_deps = ArduinoJson, AUnit
 platform = espressif32
 board = esp32dev
 framework = arduino
-build_flags = -I../test/iot/ -DUNIT_TEST
-src_filter = +<*> -<.git/> -<svn/> -<example/> -<examples/> -<bin/> -<../src/http/os/> -<_3rdParty> -<CMakeFiles>
-extra_scripts = pre:rm_submodules.py
+build_flags = -I../test/iot/ -I../test -I../src -I../src/include/cpp-client -DUNIT_TEST
+src_filter = +<*> -<.git/> -<svn/> -<example/> -<examples/> -<bin/> -<../src/http/os/> -<_3rdParty> -<CMakeFiles> -<test/lib/googletest> -<test/lib/ArduinoJson>
 upload_speed = 921600
